"ads2_m6_a1_1: 0/12 testcases passed. Cannot run hidden testcases as Server is unreachablecheck_style score: 1/1 "

+++ ADS-2-assignments/m6/ADS-2Week-1/Solution.java
@@ -6,7 +6,7 @@ class PageRank {
 	PageRank(Digraph dgra) {
 		dgraph = dgra;
 		int x = dgraph.v();
-		double y = 1 / x;
+		double y = 1.0 / (double) x;
 		System.out.println(y);
 		pR = new double[x];
         for (int i = 0; i < x; i++) {
